LF-888: drm/bridge: adv7511: Remove support for 119M mode
commit cb7340bc1e0d9e2203de54525ded19c21d9bf6ea from
https://github.com/nxp-imx/linux-imx.git

The modes requiring 119M is not a standard and tested mode by chip so, in
certain conditions, these modes are not well handled by this DSI-HDMI
converter. Since there is no detailed documentation regarding the MIPI
DPHY timings this chip supports, I'll just remove this clock from the
supported clocks.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>