MLK-23354 irqchip: gpcv2: Add wait mode workaround on imx8mp
commit 76cca923996b695a1e5fe5c9ded53e8296180338 from
https://github.com/nxp-imx/linux-imx

Add the wait mode workaround on i.MX8MP. it is just
a provisional patch for Alpha release. it will be
dropped in the future. As all the changes in this
patch need to be revered for that time, just including
all the changes of dts & driver in one patch to make
it more easier to track all the changes.

Coresight probe has some conlict with the IPI workaround.
it is meaningless to put effort on resolve such conflict,
and Coresight is not an must feature for Alpha release,
disable the Coresight support directly.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Anson Huang <Anson.Huang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>