LF-631-1 dmaengine: fsl-edma: Store and free interrupts uniformly
commit 2a563dd945ea4e50543bd32ee3c7e2388c794293 from
https://source.codeaurora.org/external/imx/linux-imx

To prepare the eDMA driver for S32V234 support, the "DMA transfer complete"
interrupt lines need to be freed (and therefore stored) in a uniform way.

After this commit, fsl_edma2_irq_init() will no longer accept an arbitrary
number of interrupts in the device tree, but only a pre-established number
of 17 interrupts, as the documentation (fsl-edma.txt) requires.

Signed-off-by: Stefan-Gabriel Mirea <stefan-gabriel.mirea@nxp.com>
Acked-by: Robin Gong <yibin.gong@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>