i2c: xiic: Enter standard mode only for > 255 byte read transfers
commit  d37b36a2a748593833048a5619cbd1256c594f64 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

When standard mode is used for all the read transfers, timing of
interrupts is not in sync with the hardware. So, to avoid breaking
existing scenarios, enter standard mode only when the size of read
transfer is > 255 bytes.

Signed-off-by: Raviteja Narayanam <raviteja.narayanam@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>