v4l: xilinx: dma: Prevent divide by zero error
commit 8e0673990aeb56b8e8e19326fd3e543fb0b84d6f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Currently there is no issue, but there is a possibility of a divide by
zero. This was found during a coverity check which gave the
following error:

"Event zero_return: Function call "lcm(dma->align, info->bpp >> 3)" returns 0."

This patch fixes this by adding a check to see if the calculated divisor
is 0.

Signed-off-by: Dylan Yip <dylan.yip@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>