v4l: xilinx: vipp: do enabling in the dependency check
commit a60caa7776a972dc61c4820348c4c5c4c987624e from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.6

This patch moves the enabling/disabling as a part of dependency check.
Since it's not just a dependency check anymore, rename the function.
It removes the error return, so there can be false postive. Not fixing
here in this commit because it'll be removed in following patch anyway.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Reviewed-by: Dylan Yip <dylan.yip@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>