clk: idt: Fix i2cwritebulk() failure
commit a672e17d692095f7967c45001785731ff6996ad1 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Commit 8b9f9d4dc511 ("regmap: verify if register is writeable before
writing operations") introduces new way for validating the writable
register range. This causes the idt driver to fail as max_register is
only 0xFF and reg address falls outside it. This works earlier as the
writeable_reg() function pointer to idt24x_regmap_is_writeable() is used
which always returns true.

So to pass the validation, the max_register is now set to 0xFFFF.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>