media: bcm2835-unicam: Use dummy buffer if none have been queued
commit  082e7986a5c3a4e14d7bb0bfe935664b05768153 from
https://github.com/raspberrypi/linux.git rpi-5.9.y

If no buffer has been queued by a userland application, we use an
internal dummy buffer for the hardware to spin in. This will allow
the driver to release the existing userland buffer back to the
application for processing.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>