MLK-23792-1: ASoC: fsl_sai: Monitor spdif rx clock in imx8mm
commit c0eca4754e8847b05a59da55735fc65337613090 from
https://github.com/nxp-imx/linux-imx.git

As we use one spare sai instance to monitor the spdif rx
clock,  there isn't belong to a sound card, we can't access
the registers by amixer controls.

So remove the amixer controls, replace them with the device
attribute.

And add an additional device attribute for enablement of
monitorring spdif. This feature only be supported on imx8mm.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Viorel Suman <viorel.suman@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>