PCI: layerscape: Change to use the DWC common link-up check function
commit 3faa46a384e9b795b5f9fa697300be8ea4367c80 from
https://source.codeaurora.org/external/imx/linux-imx

The current Layerscape PCIe driver directly uses the physical layer
LTSSM code to check the link-up state, which treats the > L0 states
as link-up. This is not correct, since there is not explicit map
between link-up state and LTSSM. So this patch changes to use the
DWC common link-up check function.

Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>