LF-1273-1 arm64: dts: imx8mp: add dsim and lcdif1 nodes
commit dbed8a17fc3384bbadc3107f3a7ebb19dc0bf191 from
https://github.com/nxp-imx/linux-imx.git

Add the 'mipi_dsi' and 'lcdif1' device nodes to i.MX8MP
platform.

Add the 'display-subsystem' node which attach to the
'lcdif_disp0' ports defined in 'lcdif1' node.

The block control module in mediamix is responsible for
clocks gating, sub-modules reset and other module control
for lcdif, dsi, csi, isp, isi and etc sub-modules which
are included in the mediamix. So create a device node to
simplify the usages.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>