MLK-23822: soc: imx: secvio: Fix boot message when nvmem not initialised
commit 5c1e731f548899a89ddde16eae792d37c28b2dbe from
https://github.com/nxp-imx/linux-imx.git

When the nvmem subsystem is not initialised at boot, the probe
will fail and an error message will be displayed.
In this case the message should not be printed as the driver will
be probed later.

This patch checks the error code from nvmem before printing the
message.

It also fixes the cleaning path as the driver was not exiting
properly.

Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>
Reviewed-by: Iuliana Prodan <iuliana.prodan@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>