media: bcm2835-unicam: Do not stop streaming in unicam_release
commit  cd9407b554e6abb3ab234246bc5f6d6dc7219577 from
https://github.com/raspberrypi/linux.git rpi-5.9.y

unicam_release calls _vb2_fop_release, which will call stop_streaming
if that particular node was streaming. Calling it unconditionally (as
the code was) means that if a second handle was opened eg to alter
a setting, on closing that connection it also stopped Unicam.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>