media: bcm2835-unicam: Do not stop streaming in unicam_release
commit  7670fbc4e104ff705ed63dda647bcc6663169226 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

unicam_release calls _vb2_fop_release, which will call stop_streaming
if that particular node was streaming. Calling it unconditionally (as
the code was) means that if a second handle was opened eg to alter
a setting, on closing that connection it also stopped Unicam.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>