MLK-24002-1 drm/imx: lcdifv3: enable shadow load for plane update
commit 6e4ceec37d692aa8d9fe76924ea12691b27b097c from
https://github.com/nxp-imx/linux-imx.git

After the atomic plane update, the shadow load should be
enabled to make sure its update can take effect on next
frame in any cases. And this enable is better to be done
in CRTC's atomic_flush() which is called after plane's
atomic_update() is called.

Besides, the shadow load enable in controller enable is
unnecessary, so remove it.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Reviewed-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>