MLK-24002-1 drm/imx: lcdifv3: enable shadow load for plane update
commit 85cabda7c471f07b2e9823265dfdf94b9c001d45 from
https://source.codeaurora.org/external/imx/linux-imx

After the atomic plane update, the shadow load should be
enabled to make sure its update can take effect on next
frame in any cases. And this enable is better to be done
in CRTC's atomic_flush() which is called after plane's
atomic_update() is called.

Besides, the shadow load enable in controller enable is
unnecessary, so remove it.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Reviewed-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>