media: bcm2835-unicam: Always service interrupts
commit  bf0e7c37986328ef54c3048a1aa487bd691ef0f8 from
https://github.com/raspberrypi/linux.git rpi-5.9.y

From when bringing up the driver, there was a check in the isr
to ignore interrupts (claiming them handled) should the driver
not be streaming.

The VPU now will not register a camera driver if it finds a
CSI2 node enabled in device tree, therefore this flawed check is
redundant.

https://github.com/raspberrypi/linux/issues/3602

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>