drm/vc4: Fix VIC usage with Broadcast RGB
commit  7d992dd984698c3e8640a79ce14eca264e39e07a from
https://github.com/raspberrypi/linux.git rpi-5.9.y

Adding the Broadcast RGB range selection broke the VIC
field of the AVI infoframes on HDMI, zeroing them for all
modes on an HDMI monitor.

Correct this so that it is only zeroed if the range is
contrary to the standard range of the mode.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>