arm64: Add workaround for Marvell erratum 38545
commit 44a6b58cb6227d2bf5056c57d8877e0eebe6eeb3 from
git@git.assembla.com:cavium/WindRiver.linux.git

On some OcteonTX2 chips, an issue exists wherein when an interrupt
is pending and a second interrupt arrives at the core under certain
stall conditions, a read of interrpt acknowledge register will return
the first interrupt but not acknowledge it. A second read of IAR will
return the first interrupt again, not the second interrupt.

Change-Id: Id5c43ff69b597fe84bc5dccdee807322305c6925
Signed-off-by: Bhaskara Budiredla <bbudiredla@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/29010
Reviewed-by: Andrew Pinski <apinski@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Tested-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
[RQ: Add cpucaps to arch/arm64/tools/cpucaps.]
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>