media: bcm2835-unicam: change minimum number of vb2_queue buffers to 1
commit b9f92b198a35215e20c06e1f4aa3b36195add6d9 from
https://github.com/raspberrypi/linux.git rpi-6.6.y

Since the unicam driver was modified to write to a dummy buffer when no
user-supplied buffer is available, it can now write to and return a
buffer even when there's only a single one. Enable this by changing the
min_buffers_needed in the vb2_queue; it will be useful for enabling
still captures without allocating more memory than absolutely necessary.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Rajeshkumar Ramasamy <rajeshkumar.ramasamy@windriver.com>