MLK-24293 gpu/imx: lcdifv3: use late system sleep pm ops
commit 1b83ae69e3f3b74528c07cc420a5784ff37e6f51 from
https://github.com/nxp-imx/linux-imx

In imx_drm_suspend() and imx_drm_resume(), the display
pipeline will be disabled and enabled respectively, so
the correct suspend sequence should be master first to
disable the pipeline, and then the components can be
suspended, and resume sequence is obviously opposite.

So, to satisfy this sequence, use the late system sleep
PM ops to replace normal sleep PM ops for this purpose.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Reviewed-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>