arm64: dts: imx8mp: Make hdmi_blk_ctrl single node for combo driver
commit 72aa6a538e7eb960e4f01f277e3e6ec55dec1c31 from
https://source.codeaurora.org/external/imx/linux-imx

Make the hdmi_blk_ctrl dts node a single one and change all the references
to it accordingly. Do this in order to have a single combo driver for
hdmi_blk_ctrl. Some of the features of the hdmi_ctrl will be used by some
different drivers in a way those drivers will know best, so adding the
syscon compatible we allow those to do just that. Those drivers will
have to also add the hdmimix PD to their devicetree node in order
to make sure hdmi_blk_ctrl is on while they are operating.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Reviewed-by: Dong Aisheng <aisheng.dong@nxp.com>
Tested-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>