gpio: siul2-s32cc: Same eirq exported by both SIULs
commit e113a47b0f964bac8e3a03a656b9648089be453d from
https://github.com/nxp-auto-linux/linux

On S32G274 same eirq is exported by two pins from different gpio chips
(SIULs).
The main issue in this case is that all related eirqs processing data
reside in SIUL2_1 no matter who gpio belongs to. To address this,
SIUL2_0 node will contain same eirqs related data as SIUL2_1
(interrupt, irq_regs).

Taking into account that same interrupt is raised no matter what
pin was configured as eirq, both gpio chips should receive the interrupt.

The first condition to achieve this is to be able to share the interrupt
between gpio chips. We remove the chained interrupt handler (as they
will overwrite each other) and request the shared interrupt using
devm_request_irq().

The second condition is to decide if an interrupt should be handled by the
current gpio chip or not. We need a "used" field to remember if the pin
belonging to the current chip is configured as eirq. If the field is set,
the interrupt should be handled.

The user shouldn't use in the same time both pins as eirq (same IMCR will
be configured when the pinmuxing is done).

Issue: ALB-4997
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>