LF-1611: arch:arm64: dts: imx8mq-evk: Fix warnings in display nodes
commit 35f0cd525fc1f99ad5aa0c4fabf9849d7baadf15 from
https://source.codeaurora.org/external/imx/linux-imx

Some of the display use-case related dts files are generating warnings
due to the incorrect usage of the 'reg' property. This patch fixes these
warnings.
Also, add the reg property to the mipi_dsi_in endpoint in dcss-adv7535.dts
(currently missing), so that the nwl-dsi driver will select the
appropriate input (DCSS).

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>