otx2-bphy-netdev: Add support for registering both rfoe and cpri netdev intf
commit 0f436120e48dc74cce103aeb1ff2164386fa5c17 from
git@git.assembla.com:cavium/WindRiver.linux.git

This patch provides support for registering netdev interface for both
rfoe and cpri links so that the interfaces are available in Linux all
the time. The Nokia interface management application is responsible
for the administrative status of each netdev interface. The resources
required for both interface types are configured during start-up
using the ODP BPHY application. The netdev driver will drop the
Tx pkts when the interface is not enabled by ODP BPHY application.

Change-Id: I83e19fedf4801c6593caecfc5fcb354cc7b1f9cc
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/31836
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>