LF-1750 gpu: imx: imx8_dprc: Get scu ipc handle in ->probe()
commit a15a9c7c66bc5ad79d67e830956f8efd773d1f1e from
https://github.com/nxp-imx/linux-imx

Instead of getting scu ipc handle from time to time in helpers
of the driver, we can get it in ->probe() once and cache it in
struct dprc.  Also, we should check the return value of
imx_scu_get_handle() to get the handle in case it returns error
code.  This makes Coverity happy.

This fixes two Coverity issues: CID 9993376 & CID 9993377.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>