x86/resctrl: Pass the required parameters into resctrl_arch_rmid_read()
commit 6a798c2d28ab2ed6e935bd6e3b3726952f602507 from
git@git.assembla.com:cavium/WindRiver.linux.git

resctrl_arch_rmid_read() is intended as the function that an
architecture agnostic resctrl filesystem driver can use to
read a value in bytes from a hardware register. Currently the function
returns the MBM values in chunks directly from hardware.

To convert this to bytes, some correction and overflow calculations
are needed. These depend on the resource and domain structures.
Overflow detection requires the old chunks value. None of this
is available to resctrl_arch_rmid_read(). MPAM requires the
resource and domain structures to find the MMIO device that holds
the registers.

Pass the resource and domain to resctrl_arch_rmid_read(). This make
rmid_dirty() too big, instead merge it with its only caller, the name is
kept as a local variable.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>