LF-1383-13 arm64: dts: freescale: Add the top level dtsi support for imx8dxl
commit 82a7c2389d0a1853c1de1ca196e11ac86a07fc4f from
https://github.com/nxp-imx/linux-imx

The i.MX8DXL is a device targeting the automotive and industrial
market segments. The flexibility of the architecture allows for
use in a wide variety of general embedded applications. The chip
is designed to achieve both high performance and low power consumption.
The chip relies on the power efficient dual (2x) Cortex-A35 cluster.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>