LF-1383-13 arm64: dts: freescale: Add the top level dtsi support for imx8dxl
commit 344820c7b86c63ac1d16c35421753537cff51b77 from
https://github.com/nxp-imx/linux-imx.git

The i.MX8DXL is a device targeting the automotive and industrial
market segments. The flexibility of the architecture allows for
use in a wide variety of general embedded applications. The chip
is designed to achieve both high performance and low power consumption.
The chip relies on the power efficient dual (2x) Cortex-A35 cluster.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>