MLK-24612-1: ASoC: fsl_esai: Add esai client driver
commit 10939ef9325a5db437a8f49f7effef85c75604f1 from
https://github.com/nxp-imx/linux-imx

ESAI client driver is working as cpu dai in front-end.
ESAI is an unity, which can't be separated to multiple device
in hardware. so we need to separate it in software view.

Each client device can accept data from user space indepedently.
Then the data of different client is mixed together before sent
to ESAI interface.

Signed-off-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Peng Zhang <peng.zhang_8@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>