irqchip/gic-v3-its: Don't return error when using mem region from atf
With the commit "drivers: gicv3: fetches atf mem region for LPI prop and
pend tables", we have added a 'prop_page' in struct rdists to distinguish
if the allocated pages are from kernel or atf, but we forgot to assign value
to it when the pages is allocated from kernel. So add the code to fix this.

After the change in commit above, its_allocate_prop_table()/its_allocate_pending_table()
will return NULL when the mem region is got from atf, so we can't consider
the return of NULL as an error, actually we should check both the return
value of its_allocate_prop_table()/its_allocate_pending_table() and the value
of gic_rdists->prop_table_va/gic_data_rdist()->pend_vaddr.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>