LF-2582-2: drm/imx/dcss: change how HDR10 LUTs/CSCs are stored
commit 3ec7629642bb8c27e3932308e299b8b8d410dae5 from
https://github.com/nxp-imx/linux-imx.git

With the latest LUT/CSC generation app changes the old way of storing the LUTs
and CSCs, as well as tagging them, proved that there could be one pipe
configuration matching 2 different tables. The result was bad colors.

The new method stores all uniq LUTs and CSCs in separate arrays and, in another
array, we hold all the different configurations with tags and indexes pointing
to the right LUT or CSC. The array is sorted by tags in order to perform a
binary search at runtime for the right entry given the input and output pipe
configurations.

Also, change the default gamut/nonlinearity for RGB formats to
Rec.2020/Rec.2084 if the color depth is 10bpc. Leave to Rec.709/Rec.709 for
8bpc formats.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>