LF-2438 arm64: kernel: smp: Add workaround for i.MX8MQ ERR11171
commit 90a0567489dd2abb188ff617eb6add499946521c from
https://source.codeaurora.org/external/imx/linux-imx

Since commit 5cebfd2d47c214 "arm64: Kill __smp_cross_call and co",
the workaround for ERR11171 for i.MX8MQ needs a different approach.
In order to do go to EL3 to power up/down the cores, right after
the ipi was raised, we call imx_gpcv2_raise_softirq which is
only built if the CONFIG_IMX_GPCV2 is enabled. Then, in
imx_gpcv2_raise_softirq depending on the existance of the ERR11171
on any platform, we call into EL3 or skip everything and get out.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>