x86/resctrl: Allow rmid allocation to be scoped by closid
commit 2c219cd9f23c0c98378bd26b8c8010a29f069819 from
git@git.assembla.com:cavium/WindRiver.linux.git

MPAMs rmid values are not unique unless the closid is considered as well.

alloc_rmid() expects the rmid to be an independent number.

Pass the closid in to alloc_rmid(). Use this to compare indexes when
allocating. If the closid is not relevant to the index, this ends up
comparing the free rmid with itself, and the first free entry will be
used. With MPAM the closid is included in the index, so this becomes a
walk of the free rmid entries, until one that matches the supplied
closid is found.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>