LF-2539-1 ARM64: dts: imx8mq-evk-dcss-rm67199: add goodix touch support
commit 4c93adfff7e97d6a89aeca08be4a9b0c74fc9368 from
https://github.com/nxp-imx/linux-imx.git

The new OLED not only use a new DSI panel(Raydium RM67199), but
also use a new touch chip, goodix gt1151q. This touch chip also
share the reset pin with the panel. According to the test, this
touch chip is not that critical with the reset operation, so
let panel control the reset pin.

Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Reviewed-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>