firmware: raspberrypi: Add support for tryonce reboot flag
commit  3956021d3a263fda759ce416b8f5243eea5033f2 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

Define a new mailbox (SET_REBOOT_FLAGS) which may be used to
pass optional flags to the Raspberry Pi firmware that changes
the behaviour of the bootloader and firmware during a reboot.

Currently this just defines the 'tryboot' flag which causes
the firmware to load tryboot.txt instead config.txt. This
alternate configuration file can be used to specify the
path of an alternate firmware and kernels allowing a fallback
mechanism to be implemented for OS upgrades.
Signed-off-by: Meng Li <Meng.Li@windriver.com>