octeontx2-af: Policy Limits for octeontx2 only
commit 4f5ce1b430a365a388b9c5945c22c8f10b01d691 from
git@git.assembla.com:cavium/WindRiver.linux.git

Policy limits will be applied only for octeontx2 family devices.
The current sysfs based method is not the ideal solution for limit
restrictions. This is not an upstreamable solution either.

The alternative method would be to use devlink. But as of today there
are limitations with the current devlink infratructure.
So resource limits are not applied for octeontx3 family.

Change-Id: Ic82ce9f6fc75469e8804efa28730ae4f134b1bda
Signed-off-by: George Cherian <george.cherian@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/38660
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Tested-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
[RQ: Move rvu_mbox_handler_free_rsrc_cnt() from rvu.c to rvu_validation.c.]
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>