spi: spi-fsl-dspi: Choose pinctrl base on pinctrl-names
commit f156b0dff00d5381a25b61026b8c523fdf81c386 from
https://github.com/nxp-auto-linux/linux

The spi controller node in dts cand define both master
and slave pinctrl. The selection between them will be done based
on pinctrl-names. The default pinctrl will be loaded first
and will be used by the master. If the controller is configured
as slave (spi-slave property is added in the dts node), the
driver will look for the "slave" pinctrl and apply it if
found.

Issue: ALB-1129
Issue: ALB-5969
Signed-off-by: Larisa Grigore <larisa.grigore@nxp.com>
Signed-off-by: Ciprian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>