clk: imx: blk_ctrl: Fix runtime PM get_sync/put balancing for resets
commit d3e9949a1f7574c2daa6e49340ee8be87cad5b0e from
https://source.codeaurora.org/external/imx/linux-imx

The pm_runtime_get_sync needs to be called only when there is no bit
asserted yet but one is about to get asserted, while the pm_runtime_put
needs to be called only after the last asserted bit gets deasserted.
Also skip the entire thing if there is a driver that mistakenly calls
the deassert for its bit first, without ever asserting it.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>