octeontx2-bphy-netdev: Add PTP hardware clock support to rfoe interfaces
commit 0b93784f73f7182de0af38c560d6fd5059703243 from
git@git.assembla.com:cavium/WindRiver.linux.git

Added PTP Hardware Cock (PHC) support to rfoe interfaces. As of now,
only the gettime64() callback is implemented by the driver as per the
requirement. The BCN offset is added to the timestamp read by gettime,
similar to the PTP timestamp calculations in the receive and transmit
path of PTP packets on rfoe interfaces.

Change-Id: If6987df64fd5efa9616b4ae3911829822e36cdfa
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/41604
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Tested-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>