firmware: xilinx: Add IOCTL support for AIE ISR Clear
commit 30ec8c8085d4a6989df0f06f31c5e5742790ca9e from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Latching of AIE NPI Interrupts is present in Versal ES1 Silicon Rev,
however it has been removed from ES2 rev.
As a result on ES1, in order to use the interrupt, a client needs to
request PMC to clear/ack the interrupt.

Provide an EEMI IOCTL to serve the same purpose. Note that, this will
only be applicable for ES1 rev. For ES2 and other non-silicon platforms,
this call will essentially be a NOP in the firmware.

Signed-off-by: Izhar Ameer Shaikh <izhar.ameer.shaikh@xilinx.com>
Signed-off-by: Wendy Liang <wendy.liang@xilinx.com>
Link: https://lore.kernel.org/r/1606722505-16194-9-git-send-email-wendy.liang@xilinx.com
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: waiting (misc)
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>