MLK-25073 arm64: dts: imx8mn-ddr3l-evk-rpmsg: reserve space for Mcore
commit 05560005716da1d9b53a3a6bca56748601c04eda from
https://source.codeaurora.org/external/imx/linux-imx

This is just a workaround for Linux 5.4 Q4 release
to avoid Linux use the memory.
This board only has 1GB memory, the 0xb8000000
exceeds the DRAM, and round back to 0x78000000,
since we not modify Mcore image, so we need to avoid
Linux touch 0x78000000 which might crash the system
and mark mcore ddr demo broken and only support booting mcore image
from U-Boot bootaux.

Reviewed-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>