drivers:fsl_qbman: remove device dependancy for link updates
commit 7c54c4506185aae831144d1bd71ed91a27fe378b from
https://source.codeaurora.org/external/imx/linux-imx

Qbman usdpaa driver is utilising the device's phy priv or net device priv
to store the eventfd context which is required to send signal to the
user space driver for link updates.
Storing the eventfd context in phy's priv is not a good idea.
In shared MAC case, kernel ethernet driver can fill phy's priv
and in that case fsl_qbman can over-write this priv, which may
cause driver mis-behaviour.

This patch introduces a new list of net devices which is
internal to fsl_qbman usdpaa driver to store the eventfd
contexts, thus removing the dependancy of device's priv.

Signed-off-by: Gagandeep Singh <g.singh@nxp.com>
DPDK-2800
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>