drivers:fsl_qbman: remove device dependancy for link updates
commit 4b9ffd77e1e100634c3d0befa8cad329020400a7 from
https://github.com/nxp-imx/linux-imx.git

Qbman usdpaa driver is utilising the device's phy priv or net device priv
to store the eventfd context which is required to send signal to the
user space driver for link updates.
Storing the eventfd context in phy's priv is not a good idea.
In shared MAC case, kernel ethernet driver can fill phy's priv
and in that case fsl_qbman can over-write this priv, which may
cause driver mis-behaviour.

This patch introduces a new list of net devices which is
internal to fsl_qbman usdpaa driver to store the eventfd
contexts, thus removing the dependancy of device's priv.

Signed-off-by: Gagandeep Singh <g.singh@nxp.com>
DPDK-2800
Signed-off-by: Meng Li <Meng.Li@windriver.com>