media: bcm2835-unicam: Return early from stop_streaming() if stopped
commit  13845e7c00d59831a1170b37ce1b93bd77421706 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

clk_disable_unprepare() is called unconditionally in stop_streaming().
This is incorrect in the cases where start_streaming() fails, and
unprepares all clocks as part of the failure cleanup. To avoid this,
ensure that clk_disable_unprepare() is only called in stop_streaming()
if the clocks are in a prepared state.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>