LF-2091 media: platform: mxc_vout: Do not check 'crop->c.width/height<0' in mxc_vidioc_s_input_crop()
commit c28802953a1e8a61987ece66f5f1e8cb7bd84fe5 from
https://source.codeaurora.org/external/imx/linux-imx

The commit f90580ca0133 ("[media] videodev2: Set vb2_rect's width
and height as unsigned") changes the width and height fields of
struct v4l2_rect from type __s32 to type __u32, which makes Coverity
complain that it's unnecessary for the platform driver to check on
the fields to ensure they are not less than zero.  This patch removes
the check to make Coverity happy.

This fixes Coverity issue: CID 18378.

Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>