LF-3299-2 regulator: add regulator driver support for FP9931
commit 07c946acf324e2f554d3672c1fd9087faa3f1e91 from
https://github.com/nxp-imx/linux-imx.git

The FP9931 includes different kinds of regulators: LDO, Boost
Converter, Buck-Boost Converter, Charge Pump and VCOM Buffer.
The typical voltages power on sequence is:

	V3P3 -> VGL -> VNEG -> VGH -> VPOS -> VCOM

The DISPLAY regulator is a virtual one which behaves like a
switch to trigger the voltages power on or power down sequence
except V3P3, since it must be switched on or off manually.

Signed-off-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Fancy Fang <chen.fang@nxp.com>
Reviewed-by: G.n. Zhou <guoniu.zhou@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>