LF-3093-2 crypto: caam - fix sm_test in case no job rings available
commit 4cf43954357201ba2ce095ff56636a735b65bfab from
https://source.codeaurora.org/external/imx/linux-imx

In case no job rings are available, secure memory structures are not
initialized, since caam_jr_probe()->init_misc_func()->caam_sm_startup()
is not called.
This could happen if all job ring devices bind to the jr uio driver
instead of the caam/jr driver.
Running sm_test in this case will lead to a crash.

Add a check to verify that SM has been initialized - if not the tests
will be skipped.

Reported-by: Gaurav Jain <gaurav.jain@nxp.com>
Signed-off-by: Horia Geantă <horia.geanta@nxp.com>
Reviewed-by: Gaurav Jain <gaurav.jain@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>