drm/vc4: hdmi: Change CSC callback prototype
commit  6ebb6ce1bb4c6b8c3061816c42e5dd677cfc2d24 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

In order to support the YUV output, we'll need the atomic state to know
what is the state of the associated property in the CSC setup callback.

Let's change the prototype of that callback to allow us to access it.

Acked-by: Thomas Zimmermann <tzimmermann@suse.de>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>