net: xilinx: Add MRMAC support
commit 146b2ba5b5d5e9798f7946ba6f04d1f37864f109 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

Add MRMAC support to AXI Ethernet driver and MCDMA.
Add support for GT reset mechanism for MRMAC subsystem to work.
This subsystem contains a common PLL reset mask register and
individual GT and data path reset bits for each lane. The GT mask
and reset bits have to be asserted in tandem whereas the data path
resets can be controlled independently. The sequence and delays
are as recommended. This patch adds support for 4x10G and 4x25G with
the former tested.

Signed-off-by: Harini Katakam <harini.katakam@xilinx.com>
Reviewed-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Signed-off-by: Sarath Babu Naidu Gaddam <sarath.babu.naidu.gaddam@xilinx.com>
Signed-off-by: Swati Agarwal <swati.agarwal@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>