dt-bindings: memory: fsl: convert ifc binding to yaml schema
commit bc899469566c0bb66884e18568544b7f36eaa3ad from
https://source.codeaurora.org/external/imx/linux-imx

Convert the txt binding to yaml format and add description.  Drop the
"simple-bus" compatible string from the example and not allowed by the
binding any more.  This will help to enforce the correct probe order
between parent device and child devices, but will require the ifc driver
to probe the child devices to work properly.

Signed-off-by: Li Yang <leoyang.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>