TEMP: dt-bindings: hwlock: Add example for AM64x SoCs
commit 6c79b7ced4d67e794aa2063cc1dba6a4cc46bd75 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Update the existing OMAP HwSpinlock binding to include an example
for AM64x SoCs. This is primarily to showcase the difference in the
IP's presence on the interconnect. This part was dropped from upstream
during reviews, but is being added back for downstream consumers.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>