net: gianfar: add ethtool eee support
commit 13704422543989252701c3e47f97e4585fe26e99 from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git

Gianfar does not support EEE, but it can connect to a PHY which supports
EEE and the PHY advertises EEE by default, and its link partner also
advertises EEE, so the PHY enters low power mode when traffic rate is low,
which causes packet loss if an application's traffic rate is low. This
patch provides .get_eee and .set_eee so that to disable the EEE
advertisement via ethtool if needed, other EEE features are not supported.

Signed-off-by: Shaohui Xie <Shaohui.Xie@nxp.com>
Signed-off-by: Yangbo Lu <Yangbo.lu@nxp.com>
Signed-off-by: Mingkai Hu <mingkai.hu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>