remoteproc: add infrastructure support for userspace driven loading
commit f84f01271e4e4deffc108645bc543ff9ced6960d from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The remoteproc infrastructure is enhanced to allow remoteproc drivers
to support userspace driven loading, with the actual boot triggered
through the kernel. This usecase is slightly different from the support
for remote processors that were already loaded and booted by bootloaders.
The support for the latter usecase is provided through RPROC_DETACHED
and new .attach() and detach() callbacks, while the userspace driven
load will usually have the remote processor firmwares loaded after
the userspace is up.

Support for the userspace loading feature is provided through a new field
in the rproc structure, 'skip_firmware_load'. This field is expected to
be set to true as per the mode/need required by the remoteproc drivers
before a rproc_add() call. The remoteproc core skips looking for firmware
and loading any firmware segments using this state flag. The default
behavior is unchanged. The sysfs 'firmware' file shows the firmware as
"unknown" and checks are added to ensure the 'auto_boot' field is not
set when the driver is configured for 'skip_firmware_load'.

The actual interface and implementation details is left to the individual
remoteproc drivers. This design will be used by the TI Keystone remoteproc
driver to support a userspace based loader. The remoteproc driver is
expected to invoke rproc_boot() and rproc_shutdown() for triggering the
boot and shutdown of the remote processor after the loading is completed
and the resource table information is published to the remoteproc driver.
The resource table is processed in-line during the rproc_boot() invocation.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>