LF-3401 spi: imx: remove CLK calculation and divider in slave mode
commit  a16312ec71605dff167a1252fd0dc3acec210d84 from
https://github.com/nxp-imx/linux-imx lf-6.1.y

In the slave mode, the clock signal is controlled by the master.
The slave does not need to calculate and configure the clock
frequency division. The slave can directly use the root clock
to sample the SCL signal.
Therefore, remove the calculation and frequency division
function of the clock in the slave mode.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>