LF-3401 spi: imx: remove CLK calculation and divider in slave mode
commit 7aea24d9af524ddb5ce08647388cdfcdc9fe86a8 from
https://source.codeaurora.org/external/imx/linux-imx

In the slave mode, the clock signal is controlled by the master.
The slave does not need to calculate and configure the clock
frequency division. The slave can directly use the root clock
to sample the SCL signal.
Therefore, remove the calculation and frequency division
function of the clock in the slave mode.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>