spi: cadence-quadspi: do not use DMA for reads smaller than 16 bytes
commit c9293d064e1103f630018efca18caf4f1ea93658 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

For reads this small it does not make a lot of sense to go through DMA
which usually has a significant setup overhead. The benefits of DMA are
reaped in larger transfers. This is especially relevant in case a
register is being polled. The DMA overhead will make the polling slower
than it should be.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>