dpaa2-eth: Set congestion group before the Rx queues become operational
commit aa6e7426c1e49c294a0cb937f835ad277079875a from
https://source.codeaurora.org/external/imx/linux-imx

Setting  the CGID for the first time or dynamically changing it
with a different one  on a FQ that has been in service (has frames in it)
is not supported, the byte/frame counts in both the old and new CGRs may be
corrupted if this is done.

This patch adds the configuration of the CGR(congestion group record)
for the network interface RX queues, in the probe function before the queues
become operational,  to prevent the corruption of the aforementioned counter.

Signed-off-by: Radu Bulie <radu-andrei.bulie@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>