irqchip/gic-v3: select and translate the partition domain
commit 337c737f6727616e89309046b4b845bb8331ae00 from
git@git.assembla.com:cavium/WindRiver.linux.git

As ACPI does not use the same scheme as DT when describing interrupts
with struct irq_fwspec, extra care is needed when translating and
matching the domain.

ACPI fwspec's put the UID of the Processor Container in param[3].
This is used directly as the part_id as Processor Container devices
are not created until the DSDT is parsed.

The extended PPI range can't be partitioned, ACPI provides the intid
directly as a GSI, whereas DT indicates the PPI number. Add a helper
fwspec_is_partitioned_ppi() to spot if this fwspec describes an
interrupt that has been partitioned.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>