irqchip/gic-v3: Print DT partitions in the same way as APCI
commit e9be2fdaa08f0e07d72e70ecfb45db0ff613a5b9 from
git@git.assembla.com:cavium/WindRiver.linux.git

DT systems print out the details of each CPU in a partition as it
is discovered. e.g.
| GICv3: GIC: PPI partition interrupt-partition-1[1] { /cpus/cpu@100[4] /cpus/cpu@101[5] }

Its going to be tricky to do this with ACPI as the PPTT helper returns
a cpumask, the DSDT parsing that would produce paths is not available
early enough. The simplest thing to do is print the set of CPUs.

Printing slightly different debug information depending on how the
platform booted is a bit naff. Change DT to just pring the set of
CPUs too. The above example becomes:
| GICv3: GIC: PPI partition: interrupt-partition-1 [1] { CPUs:<4-5> }

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>