ACPI / PPTT: Add a helper to fill a cpumask from a processor container
commit 078830bea6051a80b5ee5c3d980f19e9645d00c4 from
git@git.assembla.com:cavium/WindRiver.linux.git

The ACPI table for MPAM describes a set of CPUs with the UID of a
processor container. These exist both in the namespace and the PPTT.

Using the existing for-each helpers, provide a helper to find the
specified processor container in the PPTT, and fill a cpumask
with the CPUs that belong to it.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>