arm_mpam: Merge supported features during mpam_enable() into mpam_class
commit 439e52e8f03e253ddf2ee3d7803cb8463d1c669c from
git@git.assembla.com:cavium/WindRiver.linux.git

To make a decision about whether to expose an mpam class as
a resctrl resource we need to know its overall supported
features and properties.

Once we've probed all the resources, we can walk the tree
and produced overall values by merging the bitmaps. This
eliminates features that are only supported by some MSC
that make up a component or class.

If bitmap properties are mismatched within a component we
cannot support the mismatched feature.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>