drm: xlnx: add a new member extra_name for xlnx_bridge structure
commit 208f9ef618045b162ec04633179576a52ffec88a from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

The bridge.of_node of layer "vid" and "gfx" are the same, they both equal
to disp->dev->of_node. This will result that in the function
xlnx_bridge_debugfs_register, two layers use the same name to create
debugfs file, and there will be warning message as below when booting:

[    4.138520] debugfs: File 'xlnx_bridge-display' in directory 'xlnx-bridge' already present!

Add a new member extra_name to name to distinguish the bridges which
share the same of_node.

In addition, we need check the return value of xlnx_bridge_register.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>
Signed-off-by: Jianqiang Chen <jianqiang.chen@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>