MLK-25465-1: arm64: dts: imx8mn-evk: update vdd_soc in suspend to 0.75v
commit 827832d5f6413d99b1cd5d1f3d97222afa8b245b from
https://github.com/nxp-imx/linux-imx

To match with the latest datasheet, besides, correct vdd_soc to buck1
instead of buck2. But this patch should be no any impact since uboot
has already done the same thing. Anyway correct it, since pca9450 driver
has this ability now and could be indepedent of  uboot in the future.

Signed-off-by: Robin Gong <yibin.gong@nxp.com>
Reviewed-by: Jacky Bai <ping.bai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>