ASoC: dmaengine: Allocate buffer if substream is unmanaged
commit 5e6e33d6f2519eba610f7ab40e81f5340adf5bef from
https://github.com/linux4microchip/linux.git linux-6.6-mchp+fpga

pcm_construct/pcm_destruct callbacks are not called for BE DAIs. This means
that, if the generic dmaengine driver is used, there is no managed buffer
allocation (or pre-allocation). To be able to use the generic dmaengine
driver for BE DAI links, allocate the buffer in the hw_params callback if
there is no managed buffer.

Signed-off-by: Codrin Ciubotariu <codrin.ciubotariu@microchip.com>
[claudiu.beznea: adapt to v6.1.4]
Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>
(cherry picked from linux-6.1-trunk/at91/sound)
Signed-off-by: Andrei Simion <andrei.simion@microchip.com>
Signed-off-by: harish h <harish.h@windriver.com>