arm64: dts: ti: k3-am642-sk: Disable ICSSG nodes
commit 04abd50e251808cc22d266c5066fdc65bc0b6db4 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The ICSSG IP is not being supported as a general purpose usage on
the TI AM64x SK board, so disable the ICSSG nodes specifically for
this board. The ICSSG MDIO nodes continue to remain disabled as well,
even though the current patch makes that redundant.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>