irqchip: xilinx: Avoid __init macro usage for xilinx_intc_of_init
commit 9c0d105ad5f719c7b58646ebe594186b0c5cb7f7 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.15

This patch ensures that xilinx_intc_of_init API is not allocated into
the .init.text section. Since this API calls the API set_handle_irq
which uses __init macro to be put into .init.text section, this patch
makes changes to ensure that set_handle_irq also does not stay in
.init.text section.
This patch is in preparation for the patch through which the xilinx
intc driver will be loaded and unloaded as a module.

Existing irqchip infra does not support use cases where an irqchip
driver can be loaded as a module as part of loading of a device tree
overlay.
Experimental support for the same is being added through a irqchip
kconfig option.

Signed-off-by: Anirudha Sarangi <anirudha.sarangi@xilinx.com>
Signed-off-by: Mubin Usman Sayyed <mubin.usman.sayyed@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>