irqchip: xilinx: Add support to remove the Xilinx INTC driver module
commit 8127015e1e923fd8c4d026eab66e57003c6f9e6f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

The existing Xilinx INTC driver does not have mechanism to remove the
driver when it is being used as a module. For example, if the Xilinx
INTC IP is part of a removable fpga partition and a user intends to
replace the removable partition with a new one, the existing
implementation will not support the same.
The enhancement is done by adding a new remove API and making other
necessary changes. The added API xilinx_intc_of_remove ensures that all
interrupts already mapped to the interrupt domain and the corresponding
irq descriptors freed up.
Changes in this patch is made through a new irqchip kconfig option with the
name IRQCHIP_XILINX_INTC_MODULE_SUPPORT_EXPERIMENTAL which is disabled
by default.

Signed-off-by: Anirudha Sarangi <anirudha.sarangi@xilinx.com>
Signed-off-by: Mubin Usman Sayyed <mubin.usman.sayyed@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
State: pending
v6.1 [
Fix compilation error due to strict checking added to
IRQCHIP_MACRO by f1985002839a(irqchip: Provide stronger type
checking for IRQCHIP_MATCH/ IRQCHIP_DECLARE)
]
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>