drm/vc4: Increase the core clock based on HVS load
commit  34e970994f53b26121cd996dd07930411bd0a339 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

Depending on a given HVS output (HVS to PixelValves) and input (planes
attached to a channel) load, the HVS needs for the core clock to be
raised above its boot time default.

Failing to do so will result in a vblank timeout and a stalled display
pipeline.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>