net: enetc: manage ENETC_F_QBV in priv->active_offloads only when enabled
commit 06e362153465dc6116114156382348df97a35d8e from
https://source.codeaurora.org/external/imx/linux-imx

All the other flags in priv->active_offloads are managed dynamically,
except ENETC_F_QBV which is set statically based on the probed SI capability.

This change makes priv->active_offloads & ENETC_F_QBV track the presence
of a tc-taprio schedule on the port. Calling enetc_sched_speed_set()
from phylink_mac_link_up() is kept unconditional as long as Qbv is
supported at all on the port, because the tc-taprio offload does not
re-trigger another link mode resolve, so the scheduler needs to be
functional from the get go.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>