MLK-25558-04 clk: imx: Add the pcc reset controller support on imx8ulp
commit db5e15f47afb73b5f58d63c7e41da46007a248e5 from
https://github.com/nxp-imx/linux-imx

On i.MX8ULP, for some of the PCCs, it has a peripheral SW RST bit
resides in the same registers as the clock controller. So add this
SW RST controller support alongs with the pcc clock initialization.

the reset and clock shared the same register, to avoid  accessing
the same register by reset control and clock control concurrently,
locking is necessary, so reuse the imx_ccm_lock spinlock to simplify
the code.

Suggested-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Robin Gong <yibin.gong@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>