MLK-25558-04 clk: imx: Add the pcc reset controller support on imx8ulp
commit b347350494973326660b01ee186adeb987864285 from
https://source.codeaurora.org/external/imx/linux-imx

On i.MX8ULP, for some of the PCCs, it has a peripheral SW RST bit
resides in the same registers as the clock controller. So add this
SW RST controller support alongs with the pcc clock initialization.

the reset and clock shared the same register, to avoid  accessing
the same register by reset control and clock control concurrently,
locking is necessary, so reuse the imx_ccm_lock spinlock to simplify
the code.

Suggested-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Robin Gong <yibin.gong@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>