LF-3453 rtc:imx-sc:Handle RTC interrupt during suspend/resume
commit 77aa64cfe0d6442f1e335cf7dcf1f328693cac49 from
https://github.com/nxp-imx/linux-imx.git

Add PM notifier to imx-sc RTC driver to handle interrupt during suspend.
If an RTC interrupt occurs when the suspend process has already started
and the RTC driver has not yet suspended, the driver will request for
the suspend to be aborted.
This patch attempts to narrow the window when RTC interrupt is missed
as wakeup event. It only fixes if the interrupt occurs after the
PM_SUSPEND_PREPAREevent is issued by the kernel.

Signed-off-by: Ranjani Vaidyanathan <ranjani.vaidyanathan@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>