fpga: bridge: incorrect set to clear freeze_illegal_request register
commit 519c7cdc0b42c3545bc24db5b98efaf4f765112c from
https://github.com/altera-opensource/linux-socfpga.git socfpga-5.15.30-lts.

A Partial Region Controller can be connected to one or more
Freeze Bridge. Each Freeze Bridge has an illegal_request
bit represented in the freeze_illegal_request register.
Thus, instead of just set to clear the illegal_request bit
for first Freeze Bridge, we need to ensure the set to clear
action is applied to which ever Freeze Bridge that has
occurrence of illegal request.

Signed-off-by: Chew, Chiau Ee <chiau.ee.chew@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>