fpga: bridge: incorrect set to clear freeze_illegal_request register
commit a5ccdcf2c24654c254a768ab4e6a996e015c2a25 from
https://github.com/altera-opensource/linux-socfpga.git socfpga-6.6.22-lts.

A Partial Region Controller can be connected to one or more
Freeze Bridge. Each Freeze Bridge has an illegal_request
bit represented in the freeze_illegal_request register.
Thus, instead of just set to clear the illegal_request bit
for first Freeze Bridge, we need to ensure the set to clear
action is applied to which ever Freeze Bridge that has
occurrence of illegal request.

Signed-off-by: Chew, Chiau Ee <chiau.ee.chew@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>