LF-4069 iio: accel: mma8452: remove the reset operation during driver probe
commit 30c78dc3b9a703b075901398d9e4a5ace5707b2c from
https://source.codeaurora.org/external/imx/linux-imx

Though Sensor Datasheet define this reset bit in it's CTRL_REG2
register, but seems the actual hardware behavior do not align with
the doc expect. Once the reset bit is set, sensor can’t give back
an I2C ack, which will cause the probe fail. So just remove this
reset operation as a workaround.

Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>