rpmsg: Introduce rpmsg_create_default_ept function
commit 9a70f79825287f4af9a002fa72efff13196b9bb1 from
https://source.codeaurora.org/external/imx/linux-imx

By providing a callback in the rpmsg_driver structure, the rpmsg devices
can be probed with a default endpoint created.

In this case, it is not possible to associated to this endpoint private data
that could allow the driver to retrieve the context.

This helper function allows rpmsg drivers to create a default endpoint
on runtime with an associated private context.

For example, a driver might create a context structure on the probe and
want to provide that context as private data for the default rpmsg
callback.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>