net: hsr: Add VLAN CTAG filter support
commit 82822f92abc8c9f2215554b50678775fcbc96ef3 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This patch adds support for vlan ctag based filtering at slave devices.
The slave Ethenet device may be capable of filtering Ethernet packets
based on VLAN ID. This requires that when the VLAN interface is created
over an hsr/prp interface, it passes the vid information to the
associated slave Ethernet devices so that it updates the hardware
filters to filter Ethernet frames based on VID. This patch adds the
required functions to propagate the vid information to the slave
devices.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>