net: hsr: Save tx_flags and tag info of tagged skb
commit 51a27639d2a0a6dff49d80ba5f67f10688bf683a from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

After a tag is created for an egress skb, the tag info and slave port
info is saved in the skb's shared info area so that the saved info can
be returned to user space application by upper core layer.

In this patch, only the support of saving HSR tag info of HSR PTP
over L2 skb is added.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>