net: ti: prueth_core: hsr/prp: add HSR/PRP driver
commit d2e8eb5a46ec7216223407be3d3840648f554be0 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

HSR/PRP is a industrial Ethernet protocol defined by IEC 62439
standard. Both use 6 bytes of protocol bytes to implement redundancy.
This consists of 2 bytes of sequence number, 1 byte of Ethernet
type identifying the protocol and LSDU length and pathid/network id
bits. Link Redundancy Entity (LRE) implements 1) duplication of Ethernet
frames at Egress and 2) discard of duplicates at the ingress and 3)
processing of Supervision frames to build the Node Table information.
Node table contains the health status of remote nodes in the network.

HSR/PRP driver offload part of LRE function. In the current version of
the firmware 2 & 3 are offloaded to PRU firmware. 1) is still performed
by upper HSR/PRP driver under net/hsr. This patch initialize the firmware
based on the Ethernet type. HSR/PRP firmware uses same memory map as
Switch firmware. So all places in the driver that initializes the basic
Switch firmware is re-used for HSR/PRP as well using the
!PRUETH_IS_EMAC() macro. Any HSR/PRP specific function is initialized
under PRUETH_IS_LRE() macro.

This update support v2.1 of the ICSS firmware by using fw offsets that
are assigned at probe based on am57x platform. A list of firmware
names are also picked at probe and then based on ethtype appropriate
firmware is picked to load on the PRUs. For Packet handling, NAPI is used
at Ingress. There are two pair of Queues at the Ingress for each port.
Low priority and High priority. A NAPI instance is added for each of the
priority queue. An interrupt is raised by firmware when packets are
en-queued to one of the priority queue. red_emac_rx_packets() ensures that
packets are scanned from both queues at a specific priority and sends
them to network stack based on timestamp associated with the packet. At
the egress, packets are en-queued based on pcp bits. There are 2 queues
per port and other queues are for port to port traffic.

Firmware requires node table entries to be inserted by driver to save
PRU cycles for packet processing. Node table provides a view of remote
nodes in the network. The nodes send supervision frames periodically
with its own mac address and the same is processed by firmware. First
firmware lookup the source MAC address. If not found, it indicates a
flag in the descriptor which driver uses to insert the MAC address
in the node table. The rx packet processing code kicks a worker thread
to insert the mac address. This patch adds the required function to
do the insert function and also clear the table on demand.

Firmware also requires driver to run a timer for 10 msec to update
check flags in dram for firmware to do periodic table maintenance
functions for LRE offload.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>