v4l: subdev: add V4L2_SUBDEV_ROUTE_FL_SOURCE
commit 62e03e2f9c43d801af1d904e6bfd268131755038 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add route flag to indicate that the route is a source route. This means
that the route does not lead anywhere, and the sink_pad and sink_stream
should not be used.

A sensor which provides multiple streams should implement get_routing
and use the flag to mark the routes as sources.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>