x86/resctrl: Move get_config_index() to a header
commit 911596c755c4e6fd334e23a00d55c0d1164ec525 from
git@git.assembla.com:cavium/WindRiver.linux.git

get_config_index() is used by the architecture specific code to map a
closid+type pair to an index in the configuration arrays.

MPAM needs to do this too, there is no reason to do it differently.

Move the helper to a header file.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>