arm64: Workaround for Cortex-A53 erratum 1530924
commit 86c71e797263e11f8f69d99fada42bcb4d6c892a from
https://github.com/nxp-auto-linux/linux

Cortex-A53 erratum 1530924 allows TLB entries to be allocated as a
result of a speculative AT instruction. This may happen in the middle of
a guest world switch while the relevant VMSA configuration is in an
inconsistent state, leading to erroneous content being allocated into
TLBs.

The same workaround as is used for Cortex-A76 erratum 1165522
(WORKAROUND_SPECULATIVE_AT_VHE) can be used here.

Issue: ALB-7347
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>