MLK-25546 arm64: mm: barrier Fix 8QM usb stop when work with GPU
commit a48b76cecb1ff281eb97421b7ace207f445b266c from
https://github.com/nxp-imx/linux-imx.git

The below commit change dma_barrier from dsb to dmb.
	commit 22ec71615d824f4f11d38d0e55a88d8956b7e45f
	Author: Will Deacon <will@kernel.org>
	arm64: io: Relax implicit barriers in default I/O accessors

in 8QM, SYSBARDISABLE is 0. That's means barrier will propagate to
cci 400.

cci 400 behavior when handle barrier.

For barriers that are system shareable, that is, AxDOMAIN is 0b11,
the barrier response is sent by the CCI-400 when all pre-barrier
transactions from the barrier issuer are complete.

For other types of barrier, the response is sent when snoop responses
are received for all earlier transactions. Transactions received after
the barrier are then address-hazard-checked against all outstanding
pre-barrier transactions.

        ┌──────┐ ┌────┐
        │ A53  │ │A72 │
        └───┬──┘ └─┬──┘
            │      │
        ┌───▼──────▼──┐
        │    CCI400   │
        └─────┬───────┘
              │   1 (a)write to ddr (normal uncached memory)
              │   barrer()
              │   2 (b)write to usb register(device, nGnRE)
        ┌─────▼───────────────────────┐       ┌───────────┐
        │                             ◄───────┤   GPU     │
        │     Bus fabric              │       │           │
        └────────────────────────────┬┘       └───────────┘
3 (b) reach usb   ▲ 4 usb read   ▲   │ 6.(a)reach
         │        │   ddr        │   │
      ┌──▼────────┴─┐            │   │
      │             │            │   │
      │  USB        │      5.usb │   │
      │             │      read  │   │
      └─────────────┘            │   │
                               ┌─┴───▼─┐
                               │       │
                               │ DDR   │
                               │       │
                               └───────┘

cci 400 will response barrier (oshst) before 1(a) write to ddr.
so 3(b) reach usb before 1(a) arrive ddr. so usb get old dma descriptor,
which dma descriptor is not ready yet.

So usb hardware wait for new request. and CPU wait for usb transfer
finish.

change dmb oshst to st to let cci 400 wait for writ to ddr before
response barrier.

not impact for no-cci system because outter share is equal to system
when SYSBARDISABLE=1. NO-CCI require SYSBARDISABLE have to be 1.

kernel community discussion at
https://www.spinics.net/lists/arm-kernel/msg902879.html

Signed-off-by: Frank Li <Frank.Li@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>