phy: cdns-dphy: Add common module reset to DPHY RX
DPHY RX module has a common module reset (RSTB_CMN) among four different
other resets. The module expects all four resets to be released in a
specific order. In TI J721E SR1.0 the RSTB_CMN is internally tied to
CSI_RX_RST which is hardware controlled, so no explicit control from
software is required to reset the DPHY RX module.

With J721E SR2.0 the RSTB_CMN is separated out of the hardware controlled
and tied to a MMR bit of the DPHY RX module. An explicit write to control
the common module reset from software is required. 

Update DPHY_RX configure function to release the common module reset.
Currently this feature is specific to TI J721E SR2.0 therefore add a
compatible string and introduce a driver quirk to release the reset and
set it for J721E SR2.0

Signed-off-by: Sinthu Raja <sinthu.raja@ti.com>
[Xulin: Original patch taken from TI PROCESSOR-SDK-LINUX-J721E_08.00.00.08]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>