LF-4250 drm/bridge: nwl-dsi: Workaround RM68200 panel display shift issue
commit b3255e50de75062b50385a432e86742948117716 from
https://github.com/nxp-imx/linux-imx

Running an on-screen pxp test for about 1 hour in a while loop,
there will be horizontal display content shift observed at the
RM68200 display panel.  Display color is likely wrong at the same
time.  By tuning the HFP/HBP/HSA settings, the issue is not observed
any more.  This is based on previous knowledge got from support task
for custom external DSI bridges - change the values to be in bytes
and minus certain magic values.  Furthermore, it turns out the issue
still happens with the HSA value(9*3-10=17) as an odd number, but the
issue doesn't happen if it is rounded up to 2, thus 18, as an even
number. We are asking the design team to provide formulas to determine
the settings, but not yet available.  So, this patch implements the
workaround for now.

Reported-by: Jared Hu <jared.hu@nxp.com>
Cc: Sandor Yu <Sandor.yu@nxp.com>
Cc: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>