MGS-6371 [#imx-2489] Remove gckHARDWARE_QueryIdleUnlocked().
commit 7d1aecb353abbf5db2181569e4364d9f950005ab from
https://github.com/nxp-imx/linux-imx.git

The only place the gckHARDWARE_QueryIdle() has been used by locking the
powereMutex was in the gcvHAL_SET_FSCALE_VALUE case.
But it has been removed once with the move of QueryIdle inside of
gckCOMMAND_Stall().
Now, as there is no more usage of using it with locking, rename back the
gckHARDWARE_QueryIdleUnlocked() to gckHARDWARE_QueryIdle().

Signed-off-by: Wang Zijian <Zijian.Wang@verisilicon.com>
Signed-off-by: Nicusor Citu <nicusor.citu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>