MLK-25707-01: dma: pxp_v3: add software restart for iMX8ULP
commit 6043bdddc2f19d11c6ea68714b912a086563ba07 from
https://source.codeaurora.org/external/imx/linux-imx

When pxp driver probe, it will do software reset and reconfigure
operation. For iMX8ULP, power domain will be turned off in runtime
suspend and the status will be missed, so need to do software reset
and reconfigure operation again. But for other platform, such as 7D,
it's not necessary.

Signed-off-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>