MLK-25707-01: dma: pxp_v3: add software restart for iMX8ULP
commit 608692314dfa19006d716c575535fd1f2b523014 from
https://github.com/nxp-imx/linux-imx.git

When pxp driver probe, it will do software reset and reconfigure
operation. For iMX8ULP, power domain will be turned off in runtime
suspend and the status will be missed, so need to do software reset
and reconfigure operation again. But for other platform, such as 7D,
it's not necessary.

Signed-off-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>