LF-3880: ASoC: SOF: Fix compress stream cleanup path
commit d7f2835b39db83304731904aab72d8216ea70b3b from
https://source.codeaurora.org/external/imx/linux-imx

Consecutive runs of cplay now fails because after first
cplay the Firmware components states is not reset.

We need to send SOF_IPC_STREAM_PCM_FREE command when a stream is
closed in order to cleanup components states in Firmware.

This will cause a pipeline reset and it will put the pipeline +
components in a known state.

Reviewed-by: Peng Zhang <peng.zhang_8@nxp.com>
Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>