arch/arm/mach-axxia: use mpidr_to_hwcpu to retrieve real hardware cpuid
commit df012ccaa8d74dc56bea166a913e00a0fabc322f from
https://github.com/axxia/linux-yocto.git v5.10/standard/axxia-dev/base.

In function arm_dt_init_cpu_maps, it will use hwid whose value is "reg" property
in cpu node to compare with mpidr whose value read from MPIDR register. When boot
in normal condition, hwid is equal to MPIDR value which is "0" by chance. But in
kdump, the boot cpu will be any of 16 cpus and arm_dt_init_cpu_maps will not be able
to get correct hardware cpuid and this will result deviant irq affinity setting.

So at first, change "reg" property to mpidr value, and then use mpidr_to_hwcpu to transit
it to get real cpuid. Or else, in kdump, most irq will not be triggered.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>