drm/vc4: hdmi: Warn if we access the controller while disabled
commit  680639f05767778794628470cfde561b550ad755 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

We've had many silent hangs where the kernel would look like it just
stalled due to the access to one of the HDMI registers while the
controller was disabled.

Add a warning if we're about to do that so that it's at least not silent
anymore.

Reviewed-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Link: https://patchwork.freedesktop.org/patch/msgid/20210819135931.895976-7-maxime@cerno.tech
Signed-off-by: Meng Li <Meng.Li@windriver.com>