LF-4316 drm/panel: panel-raydium-rm67191: Set DSI mode to 0x0B for RM67199
commit 671757b4afb52ea57039378e080af9b596f74456 from
https://github.com/nxp-imx/linux-imx.git

Without this patch, the driver sets the DSI mode to 0x08 for RM67199,
which means 'command RAM mode' as the panel data sheet indicates.
In that mode, something like screen tearing effect can be seen when
doing page flip test by using modetest.  This patch changes the DSI
mode to 0x0B for RM67199, which means 'video RAM capture mode' and
aligns to the setting of RM67191.

Fixes: ac708b5b2804 ("MLK-24805-1: drm/panel: rm67191: Add support for rm67199")
Cc: Sandor Yu <Sandor.yu@nxp.com>
Cc: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>