LF-4502 drm/imx: dpu: dpu-crc: Avoid out-of-bounds access to array 'params' for ROI
commit 6728e190316056e0434fbce5bbf4e5f729ddeaf6 from
https://github.com/nxp-imx/linux-imx.git

The number of Region Of Interest(ROI) parameters(delimited by comma) is 4,
so we should stop getting parameters if the iterator i is greater than 3
where i starts from 0.  We cache the parameters in the array 'params' whose
size is also 4.  This patch fixes the check for the iterator to avoid
out-of-bounds access to the array 'params'.

This fixes Coverity issue: CID 12358615.

Cc: Sandor Yu <Sandor.yu@nxp.com>
Cc: Laurentiu Palcu <laurentiu.palcu@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Liu Ying <victor.liu@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>