LF-4506-2: drm: bridge: cdns_hdmi: save new connector state
commit 8616685f93a86fb0bb701efe65d4967a65786e0a from
https://github.com/nxp-imx/linux-imx.git

Colorspace connector property is use to choose the right output colorspace
in mode_fixup.
The new colorspace state is not stored to the current state pointer until in atomic_commit.
The current state pointer is not updated and incorrect output colorspace
will be selected in mode_fixup.

Save new connector state to struct cdns_mhdp_device and use the new
state to select correct output colorspace.

Signed-off-by: Sandor Yu <Sandor.yu@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>