arm_mpam: Add debugfs entries to show the MSC/RIS the driver discovered
commit b1ea5617320f9fb71383af6058b653c11d116e4c from
git@git.assembla.com:cavium/WindRiver.linux.git

Not all of MPAM is visible through the resctrl user-space interface.
To make it easy to debug why certain devices were not exposed through
resctrl, allow the properties of the devices to be read through debugfs.

This adds an mpam directory to debugfs, and exposes the devices as well
as the hierarchy that was built.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>