LF-4336-2 PCI: imx: to solve the compatible issue do perst properly
commit 841bf366158fe120f871c8e919251deabf8c4f05 from
https://source.codeaurora.org/external/imx/linux-imx

PCIe r4.0 sec 6.6.1, Conventional Reset.
The set of rules addresses requirements placed on the system:
To allow components to perform internal initialization, system
software must wait a specified minimum period following the
end of a Conventional Reset before it is permitted to issue
Configuration Request to the device immediately below that
Port.
- With a Downstream Port that does not support Link speeds greater than
  5.0 GT/s, software must wait a minimum of 100 ms before sending a
  Configuration Request to the device immediately below that Port.
- With a Downstream Port that supports Link speeds greater than
  5.0 GT/s, software must wait a minimum of 100 ms after Link training
  completes before sending a Configuration Request to the device
  immediately below that Port.

Unfortunately, ther is a compatible link down issue when NEC Corporation
uPD720200 USB 3.0 Host Controller (rev 04) is used on i.MX7D SDB board.
When add the 100ms delay just at the end of the reset.

To level up the compatibility and align with the SPEC, add the 100ms
wait after the end of the PERST# toggle and link is up to fix the link
down compatible issue when NEC uPD720200 is used on i.MX7D SDB board.

And refer to PCIe r4.0 sec 6.6.1, A component must enter the LTSSM
Detect state within 20 ms of the end of Fundamental Reset.

To level up the compatibility of i.MX8MP PCIe to support one PLEXTOR
NVME SSD device. Add one 20ms delay at the end of the PERST# toggle and
before the start of the link training for i.MX8MP PCIe.

Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>
Reviewed-by: Jun Li <jun.li@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>