arm_mpam: Track bandwidth counter state for overflow and power management
commit 7346a0c47546f1526c1dc3a34891cee28eeb943c from
git@git.assembla.com:cavium/WindRiver.linux.git

Bandwidth counters need to run continuously to correctly reflect the
bandwidth.

The value read may be lower than the previous value read in the case
of overflow and when the hardware is reset due to CPU hotplug.

Add struct mbwu_state to track the bandwidth counter to allow overflow
and power management to be handled.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>