arm64: dts: ti: k3-am65: Fix the clock order and message ram configuration
commit 416a77dd27f1702cf2835bd8ccad274e264aec7c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

As per the dt-binding of bosch mcan, the clock order should be hclk and
then cclk. So, fix the clock order in the mcan device tree nodes and
allocate all the available resources in the message ram.

Also, when there is no reg property in a dt node, the node names should not
have a '@' symbol. Therefore, remove it from node names of the
transceivers.

Fixes: fd8f83bb2741 ("arm64: dts: ti: k3-am65-mcu: Add Support for MCAN")
Fixes: 4c1c889dafe7 ("arm64: dts: k3-am654-idk: Add Support for MCAN")
Signed-off-by: Aswath Govindraju <a-govindraju@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>