arm64: dts: ti: k3-j721e: Fix the clock order and message ram configuration
commit 892ae8ead814289a53cbae8cae44750cda5a6268 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

As per the dt-binding of bosch mcan, the clock order should be hclk and
then cclk. So, fix the clock order in the mcan device tree nodes and
allocate all the available resources in the message ram.

Also, when there is no reg property in a dt node, the node names should not
have a '@' symbol. Therefore, remove it from node names of the
transceivers.

Fixes: e23348910969 ("arm64: dts: ti: k3-j721e: Add support for MCAN nodes")
Fixes: f4a5e8c093fb ("arm64: dts: ti: k3-j721e-common-proc-board: Add support for mcu_mcan nodes")
Signed-off-by: Aswath Govindraju <a-govindraju@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>