arm_mpam: Add helper to reset saved mbwu state
commit 3f86672c7e16489aab966f83dfd6a7060b44b7d7 from
git@git.assembla.com:cavium/WindRiver.linux.git

resctrl expects to reset the bandwidth counters when the filesystem
is mounted.

To allow this, add a helper that clears the saved mbwu state. Instead
of cross calling to each CPU that can access the component MSC to
write to the counter, set a flag that causes it to be zero'd on the
the next read. This is easily done by forcing a configuration update.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>