LF-3469: net: phy: at803x: Disable PHY hibernation during init
commit c437d2b50ad467e82c754a0cc2016f86dde09677 from
https://source.codeaurora.org/external/imx/linux-imx

The Atheros 803x PHY will go to hibernate mode after
10 seconds if no activity on the link.
When in hibernation, it will not provide any clock to the MAC.

This caused issue when trying to bring up the interface when
no cable was connected: MAC driver would timeout, and the PHY
power domain would stay on. It is also possible that this caused
issues with EEE capable remote PHY.

Disabling this feature during initialization to avoid potential
side effects.

Signed-off-by: Seb Haezebrouck <sebastien.haezebrouck@nxp.com>
Reviewed-by: Shenwei Wang <shenwei.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>