arm: dts: dra72: Add VIP mux pinctrl node and bindings
commit b4e97ce3e93b835d282c1dd1e33105cdf9f3c2e8 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

DRA72 device has only one instance of VIP (VIP1)
But the SoC has more video pads(total 8) than available ports
(1a, 2a, 1b, 2b). Each video port can be connected to one of
the pad group. This is selected via a control module register
CTRL_CORE_VIP_MUX_SELECT.

Depending on the board design, some of the interfaces would be used
and this register needs to be programmed.
As this can be best fit as a pinctrl device, Add a pinctrl node in
dra72x.dtsi and also add the bindings for the bitfields to be used
when configuring this mux.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>
Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>