x86/resctrl: Add stubs for the IOMMU helpers
commit 7466a3e2478a4620ccb04e757f22bf09a364279e from
git@git.assembla.com:cavium/WindRiver.linux.git

Arm's MPAM has support for assigning devices behind an IOMMU to a
control or monitor group. Helpers are needed because of the way
MPAM emulates CDP.

Add stub versions of the arch helpers used to do this. This allows
the code to depend on IS_ENABLED(), allowing the compiler to check
that code on platforms that don't enable the feature.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>