MXEM-88 arm64: dts: imx8mq: disable USB PHY rxdect at P3
commit b54aee34e60f4476939c48d62292f0ae2788555f from
https://github.com/nxp-imx/linux-imx

Cusomter found one USB host issue at low temmperature(-25),
USB host can not maintain the U3 link state with a USB3 Hub,
this issue turns out to be caused by one known issue of Synopsys
USB3.0 PHY used in iMX8MQ, errata doc description directly
quoted as below:

1. Purpose
A potential issue has been found in the Host USB3.0 PHY product, This
issue does not impact the USB3.0 PHY device operation. This document
will describe the circumstances under which this issue can occur, the
effects of this issue, and any workaround as appropriate.

2. Circumstances of Occurrence
The issue could be seen in Host application when the sequence below
occurs:
1) The Synopsys PHY Host performs an RX.Detect operation in the P3 state
2) The Synopsys PHY TX Host output that is disabled is above common
mode.
This issue is only seen in an extreme corner case of the process where
the PHY TX output when in P3 and not holding common mode floats
substantially above common mode.

3. Issue Descriptions
In P3 the TX output is not specified or designed to hold a common mode
voltage. The output voltage typically naturally discharges to below
common mode, if not the RX detect pulse generated might not meet the
specification and can cause the RX.Detect function to incorrectly
determine a receiver is not present.
The sequence of the failure is as follows:
1) The Synopsys PHY is in P3
2) During the time the common mode is disabled, the TX output rises
above common mode.
3) An RX.Detect operation is initiated.
4) The Detect pulse amplitude is not in the specification range.
5) The RX.Detect result is that a receiver is not present.

4. Consequences of the Issue
Depending on the state of the LTSSM, the USB3.0 Phy will either fail to
detect a connected SS receiver and revert to HS mode or the link partner
will go to compliance.

5. Workaround
Set controller bit GUSB3PIPECTRL[28] = 1 (Perform receiver detect in P2)
to allow the PHY to establish common mode before requesting a receiver
detect.

6. Consequence of the Workaround
If using the optional hibernation feature, it must be disabled if
GUSB3PIPECTRL[28] = 1. As imx8MQ does not has hibernation feature,
so there is no impact.

iMX8MP use a newer version USB3.0 PHY with this issue fixed, so no need
this workaround.

Reviewed-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Li Jun <jun.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>