include: linux: firmware: Add header file with error events details
commit e66cf69310cc7ccb4141fc0902a078cfc6dfa67f from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.1

Order of this patch need to be changed. Header to be added before user.

Add new header file xlnx-error-events.h to use with xilinx event management
driver.
This file contains the MACROs for Error Event Node_Id and Error Event Mask
values. From this header file agent also able get the information on the
combination of Node_Id and Mask which need to be use during registration
of any error event.

Signed-off-by: Abhyuday Godhasara <abhyuday.godhasara@xilinx.com>
Signed-off-by: Rajan Vaja <rajan.vaja@xilinx.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>