LF-4675 regulator: fp9931: Defer probe when retrieve gpio which is not ready
commit c4274e4255ec8ed94dcaa591231cdfa66ff6e670 from
https://github.com/nxp-imx/linux-imx.git

The callback trace:

fp9931_regulator_probe
  -> fp9931_pmic_dt_parse_pdata
    -> of_get_named_gpio

In fp9931_pmic_dt_parse_pdata function, it returns -EINVAL (-22) when gpio is
not ready which caused probe failure as following messages. The right way to
respond to gpio not-ready case should pass return value of of_get_named_gpio
to the caller and then do defer probe for pmic driver. This patch fixed it.

[    4.078902] fp9931-pmic fp9931-pmic: no epdc pmic wakeup pin available
[    4.085496] fp9931-pmic: probe of fp9931-pmic failed with error -22

Signed-off-by: Robby Cai <robby.cai@nxp.com>
Reviewed-by: Dong Aisheng <aisheng.dong@nxp.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>