net: ti: prueth: hsr: add support for packets timestamping
commit 0dc2edeb42b305ab613187bef89d11f8b167d551 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When PTP support is enabled in hsr firmware, the firmware timestamps the
ptp packets that are going out via prueth ports. The timestamp value is
copied to a shared memory location. After timestamping, a host irq is
raised by the firmware as a notification to read the timestamp.

When a HSR non-link-local PTP event is cut-through forward by the HSR
firmware, in addition to the rx timestamp, a tx timestamp is also provided
by the firmware. The tx timestamp is saved in the shared redundant info
area of the received skb before the skb is passed to upper layer. Currently
cut-through tx timestamp is handled only for HSR PTP over L2 non-link-local
event.

Add support for handling this ptp tx irq, read the tx timestamp and pass it
to userspace layer. In case, of cut through - on RX get RX timestamp and
delay PTP pcaket delivery until TX timestamp is ready.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>