cacheinfo: Set cache 'id' based on DT data
commit 5a68c8abcf22cb31519e3cd1a733dadc6a384bf3 from
git@git.assembla.com:cavium/WindRiver.linux.git

Use the minimum CPU h/w id of the CPUs associated with the cache for the
cache 'id'. This will provide a stable id value for a given system. As
we need to check all possible CPUs, we can't use the shared_cpu_map
which is just online CPUs. As there's not a cache to CPUs mapping in DT,
we have to walk all CPU nodes and then walk cache levels.

Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: "Rafael J. Wysocki" <rafael@kernel.org>
Signed-off-by: Rob Herring <robh@kernel.org>
Signed-off-by: James Morse <james.morse@arm.com>
[RQ: Replace of_get_cpu_hwid() with of_cpu_node_to_id().]
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>