media: cadence: csi2rx: Use link freq when possible
commit b53bdec06081f87c4203427ae94e10c63e57c7bc from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

In a multiplexed stream setup, the pixel rate is not the correct metric
to calculate link frequency. It is likely that the source won't even
implement the control. The link frequency control is a better
replacement for it. Use v4l2_get_link_freq() that abstracts both the
pixel rate and link frequency controls.

Use frame descriptors if available to calculate bpp. This can also be
used as an indication that source is sending a multiplexed stream, in
which case there should be no bpp.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>