LF-4558-3 gpio: gpio-imx-rpmsg: do not alloc msg dynamically
commit 853a3c8e3bbae9d7011e4591996d667a931eb68d from
https://source.codeaurora.org/external/imx/linux-imx

We meet system random hang when do system suspend/resume. This is
because we use a work queue to send the rpmsg info, before this
queue really do the msg work, driver already free the msg.

And also there are some cases when in irq context, we need to use
this msg, if use zalloc() to request memory dynamically, there is
a risk because zalloc may sleep.

So here change to use msg directly to struct, use static method
to avoid the upper two issues.

Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>