cacheinfo: Expose the code to generate a cache-id from a device_node
commit 9f7b0dc36fd5bf653fcc166ce721a19415832d0c from
git@git.assembla.com:cavium/WindRiver.linux.git

The MPAM driver identifies caches by id for use with resctrl. It
needs to know the cache-id when probeing, but the value isn't set
in cacheinfo until device_initcall().

Expose the code that generates the cache-id. The parts of the MPAM
driver that run early can use this to set up the resctrl structures.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>