ARM: dts: am57xx: Add PRU system events for virtio
commit 9aaf81c21d293e28e3f2752611ba8567a79d928c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

A PRU system event "vring" has been added to each of the PRU nodes in
the PRU-ICSS remote processor subsystem to enable the virtio/rpmsg
communication between MPU and that PRU core. The additions are done
in the common am57-pruss.dtsi, and so are inherited by all the AM57xx
boards.

The PRU system events is the preferred approach over using OMAP
mailboxes, as it eliminates an external peripheral access from
the PRU-side, and keeps the interrupt generation internal to the
PRUSS. The difference from MPU would be minimal in using one
versus the other.

Mailboxes can still be used if desired, but currently there is no
support on firmware-side for the SoC to use mailboxes. Either approach
would require that an appropriate firmware image is loaded/booted on
the PRU.

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>