media: bcm2835-codec: Limit video callbacks
commit  47d5c7d1372a84fc2283f3afa05620d248372fb3 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

Limit the number of allowed video callbacks. This helps with limiting
the size of the coded input FIFO which in turn helps to control latency.
Choose -5 as the magic number as it translates to DPB+5 buffers which
has been proven to be a good number in the past.

Ideally coded buffers would not be returned to the user until they
had been decoded into the DPB or been discarded as bad, but that grade
of control is unavailable to us.

Signed-off-by: John Cox <jc@kynesim.co.uk>
Signed-off-by: Meng Li <Meng.Li@windriver.com>