x86/resctrl: Only allocate mbps_val[] when it is going to be used
commit 7f580d4ebb23ed3ab980ab40e85e8b2ac6d7193a from
git@git.assembla.com:cavium/WindRiver.linux.git

mbps_val[] is allocated by resctrl to store the mba_sc configuration
value when that mode is enabled.

The array is allocated whenever a domain is online, regardless of
whether mba_sc is in use, or whether the filesystem is mounted.

Move the allocation so that the memory is only allocated when it is
going to be used. Check is_mba_sc() when the domain is brought online,
this can only be true while the filesystem is both mounted, and mba_sc
is enabled. To cover domains that are already online when the filesystem
is mounted, mba_sc_allocate() allocates the mbps_val[] for each domain
from set_mba_sc().
Similarly, set_mba_sc() is called to disable mba_sc, this call can
free the mbps_val[] array.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>