drm/vc4: hdmi: Simplify the connector state retrieval
commit  631a7e52cfcafb43e5d79a51994b755023270635 from
https://github.com/raspberrypi/linux.git rpi-5.15.y

When we have the entire DRM state, retrieving the connector state only
requires the drm_connector pointer. Fortunately for us, we have
allocated it as a part of the vc4_hdmi structure, so we can retrieve get
a pointer by simply accessing our field in that structure.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>