arm_mpam: Allow the maximum partid to be overridden from the command line
commit 26774c4a9af5b9b1acdbdda4f63006030ac04a31 from
git@git.assembla.com:cavium/WindRiver.linux.git

MPAMs bandwidth monitors are only available via resctrl if there are
enough monitors for each combination of partid and pmg to have one.

As it is unlikely anyone built that many monitors, allow the
maximum partid the system will use to be set from the kernel
command-line.

With this, it should be possible to bandwidth monitors to be
enabled by reducing the number of partid in use.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>