HSD #1308655004: don't clear the ue/ce status bits
commit e059ee9be7a270223d9368c2c95ac3f6cb44d638 from
https://github.com/altera-opensource/linux-socfpga.git socfpga-5.15.30-lts.

When there is a uncorrected or double-bit error, the kernel will panic
and will result in a watchdog reset. When the SDM gets the reset, it
needs to know that a DBE had occurred so it can scrub the DDR on a
reboot. Thus we should not clear the CE/UE status bits in the EDAC.

Signed-off-by: Dinh Nguyen <dinh.nguyen@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>