HSD #1308655004: don't clear the ue/ce status bits
commit 2fbc58603c1cfc974383919a42e7e1ff79a24313 from
https://github.com/altera-opensource/linux-socfpga.git

When there is a uncorrected or double-bit error, the kernel will panic
and will result in a watchdog reset. When the SDM gets the reset, it
needs to know that a DBE had occurred so it can scrub the DDR on a
reboot. Thus we should not clear the CE/UE status bits in the EDAC.

Signed-off-by: Dinh Nguyen <dinh.nguyen@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>