octeontx2-nicvf: Free VF PTP resources.
commit b9bf71b021099366ead230eff9e6e6226209c15e from
git@git.assembla.com:cavium/WindRiver.linux.git

When a VF is removed, respective PTP resources are not uninitialized.
Because of this PTP resources may still be under use even after deleting
VF, which may result to "NULL pointer dereference" and segfault of the
application which uses these resources. This patch fixes this by
uninitializing the PTP resources when VF is removed and during failure
cases of VF probe.

Fixes: 0bbd6bd70d4b ("octeontx2-nicvf: Add PTP hardware clock support to
		     NIX VF")
Change-Id: I7645a77346307571ad7d9e85ddec516c826b7925
Signed-off-by: Rakesh Babu Saladi <rsaladi2@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/65245
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>